pyproject.toml
[tool.poetry]
name = "tft-glucose"
version = "0.1.0"
description = "Temporal Fusion Transformer for Blood Glucose Prediction"
authors = ["Your Name <your.email@example.com>"]
readme = "readme.md"
packages = [{include = "models"}]

[tool.poetry.dependencies]
python = "^3.11"
torch = {version = "2.5.1+cu121", source = "pytorch"}
torchvision = {version = "0.20.1+cu121", source = "pytorch"}
torchaudio = {version = "2.5.1+cu121", source = "pytorch"}
pytorch-lightning = ">=1.5.0,<2.5.3"
darts = "^0.37.1"
seaborn = "^0.13.2"
tabulate = "^0.9.0"
chronos-forecasting = "^1.5.3"
pandas = "^2.3.2"
numpy = "^2.3.3"
matplotlib = "^3.10.6"
scikit-learn = "^1.7.2"
optuna = "^4.5.0"
tensorboard = "^2.20.0"
tqdm = "^4.67.1"
statsmodels = "^0.14.5"
jupyter = "^1.1.1"
ipykernel = "^6.30.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
black = "^22.0.0"
isort = "^5.10.0"
flake8 = "^5.0.0"

[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu121"
priority = "explicit"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
